map divided into links to sectors
