+++ src/pages/add_sensor/add_sensor.js
@@ -2,7 +2,6 @@ import * as React from "react";
 import { Component } from "react";
 import "../../style/global.css";
 import "../../style/add_sensor.css";
-import { Route, BrowserRouter as Router, Switch, Link } from "react-router-dom";
 
 class AddSensor extends Component{
     constructor(){
