+++ src/pages/index.js
@@ -4,8 +4,9 @@ import Map from "./map/map";
 import MapSidebar from "./map/map_sidebar";
 import "../style/global.css";
 import "../style/mainpage.css";
-import { Route, BrowserRouter as Router, Switch, Link } from "react-router-dom";
-import Sector from "./sector/sector";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { Link } from "gatsby";
+import Sector from "./sector/sector_details";
 
 class App extends Component { // TODO: put this in App.js
 	constructor() {
@@ -16,16 +17,17 @@ class App extends Component { // TODO: put this in App.js
 		return(
             <Router>
 				<div className="mainpageContainer">
-					<Switch>
+					{/* <Switch>
 						<Route exact path="/">
-							<Link to={"/sector/sector"}>
+							<Link to={"/sector/sector_details"}>
 								<Map className="map"/>
 							</Link>
 						</Route>
 						<Route exact path="/sector/sector">
 							<Sector className="sector"/>
 						</Route>
-					</Switch>
+					</Switch> */}
+					<Map className="map"/>
 					<MapSidebar className="mapSidebar"></MapSidebar>
             	</div>
 			</Router>
