+++ src/style/map.css
@@ -1,4 +1,33 @@
+.mapContainer{
+    display: grid;
+    grid-template: 1fr / 1fr;
+}
+
 .mapImg{
+    z-index: 1;
     width: 75vw;
     height: 100vh;
+    grid-column: 1 / 1;
+    grid-row: 1 / 1;
+}
+
+.sectorsContainer{
+    z-index: 2;
+    width: 75vw;
+    height: 100vh;
+    grid-column: 1 / 1;
+    grid-row: 1 / 1;
+    display: grid;
+    grid-template-rows: repeat(100, 1fr);
+    grid-template-columns: repeat(100, 1fr)
+}
+
+.sector{
+    margin: 0;
+    z-index: 3;
+    background-color: red;
+    opacity: 20%;
+    width: 100% !important;
+    height: 100% !important;
+    border: 1px black solid;
 }
