+++ src/style/sector_details.css
